chore(team): split mailbox service rpc
