+++ src/team/manager.rs
@@ -14,6 +14,7 @@ mod mailbox_queries;
 mod mailbox_reply_obligations;
 mod mailbox_service;
 mod mailbox_service_channels;
+mod mailbox_service_rpc;
 mod mailbox_shared_thread;
 mod mailbox_sqlite;
 mod mailbox_store;
